Leafes_ready_realization/requirements.txt
numpy
Pillow==10.3.0
pyTelegramBotAPI==4.17.0
scipy==1.13.0
torch==2.2.1
torchvision==0.17.1
timm
opencv-python
aiohttp